fix: manually check versions of dependencies #153
The issue is that all Redmine plugins are essentially combined into a single monolith. When installing dependencies, Bundler does not just consider a specific Gemfile, but everything in the monolith, including Redmine's Gemfile. This can lead to conflicts if the same Gem is present in both our Gemfile and someone else's. Unfortunately, we do not have many options other than omitting version specifications for our dependencies. To make this less disheartening, we can manually check if the required versions are available. If they are missing, we can notify the user and hope that our plugin will still work with their installed dependencies.